source-code/backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.7.0
dj-rest-auth==7.0.1
django-allauth==65.4.1
requests==2.32.3
djangorestframework_simplejwt==5.5.0
